+++ CHANGELOG.md
@@ -17,4 +17,5 @@ BUG FIXES:
 DEPRECATIONS:
 
 * resource/sumologic_collector_ingest_budget_assignment: Deprecated in favor of assigning ingest budgets through the _budget field attribute of collectors [GH-135]
-* resource/sumologic_collector: Deprecated `lookup_by_name` and `destroy` attributes
+* resource/sumologic_collector: Deprecated `lookup_by_name` and `destroy` attributes [GH-136]
+* resource/sumologic_sources: Deprecated `lookup_by_name` and `destroy` attributes [GH-137]